chore(asm): move weak algorithms fixtures
