+++ tests/appsec/iast/fixtures/weak_algorithms.py
